% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colLogSumExps}
\alias{colLogSumExps}
\title{Stable computation of the logarithm of a sum for columns of a matrix}
\usage{
colLogSumExps(logV, accumulators = 5L)
}
\arguments{
\item{logV}{The input vector.}

\item{accumulators}{The number of accumulators used (maximum 12).
Typically 5 is fastest here, but this may be architecture-dependent.
For values in the range [2,12], the difference in speed is typically
not large (less than 10%).}
}
\description{
Numerically stable computation of
  \code{apply(X,2,function(x) log(sum(exp(x))))}
using SIMD instructions. Note that no function \code{rowLogSumExps}
is provided. If this is desired, the matrix should first be transposed
and then can be passed into \code{colLogSumExps}.
}
\details{
See \url{https://github.com/1j/logSumExp}
}
\examples{
logx = matrix(runif(1e6,-10000,-9000),1000,1000)
logsum = colLogSumExps(logx)
}
