% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logAddExp}
\alias{logAddExp}
\title{Stable computation of element-wise sum of two vectors in log space}
\usage{
logAddExp(logA, logB)
}
\arguments{
\item{logA}{The first input vector.}

\item{logB}{The second input vector.}
}
\description{
Numerically stable computation of
  \code{log(exp(logA)+exp(logB))}
using SIMD instructions.
}
\details{
See \url{https://github.com/1j/logSumExp}
}
\examples{
## create two vectors of similarly small log values
logA = runif(1e5,-10000,-9000)
logB = logA + rnorm(1e5,0,1)

## compute sum in log space
logC = logAddExp(logA,logB)
head(cbind(logA,logB,logC))
##           logA      logB      logC
## [1,] -9782.350 -9781.286 -9780.990
## [2,] -9279.501 -9279.381 -9278.746
## [3,] -9402.641 -9402.887 -9402.064
## [4,] -9927.700 -9926.675 -9926.369
## [5,] -9586.167 -9586.246 -9585.513
## [6,] -9602.263 -9603.234 -9601.942
}
