% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logSumExp}
\alias{logSumExp}
\title{Stable computation of the logarithm of a sum of values}
\usage{
logSumExp(logV, accumulators = 8L)
}
\arguments{
\item{logV}{The input vector.}

\item{accumulators}{The number of accumulators used (maximum 12).
Typically 8 is fastest, but this may be architecture-dependent.
For values in the range [2,12], the difference in speed is typically
not large (less than 10%).}
}
\description{
Numerically stable computation of
  \code{log(sum(exp(logV)))}
using SIMD instructions.
}
\details{
See \url{https://github.com/1j/logSumExp}
}
\examples{
logx = runif(1e5,-10000,-9000)
log(sum(exp(logx)))
## [1] -Inf
logSumExp(logx)
## [1] -8995.453

## Example with a normalized vector
x = runif(1e5)
x = x/sum(x)
logSumExp::logSumExp(log(x))
## [1] 0
}
